フォローボタンのコンポーネント化
