+++ components/feeds/list/FeedItem.tsx
@@ -1,13 +1,8 @@
 'use client'
 import { FaUser } from 'react-icons/fa'
-import { createFollowFeed, deleteFollowFeed } from '@/app/feeds/list/_action/actions'
-import { followingFeedAtom } from '@/jotai/followingFeedAtom'
 import { Feed } from '@prisma/client'
-import { useAtom } from 'jotai'
-import { useSession } from 'next-auth/react'
 import Link from 'next/link'
-import { useRouter } from 'next/navigation'
-import { useEffect, useState } from 'react'
+import FollowFeedButton from './FollowFeedButton'
 
 const FeedItem = ({
   item
@@ -18,37 +13,6 @@ const FeedItem = ({
     }
   }
 }) => {
-  const [isFollowed, setIsFollowed] = useState(false)
-  const [followingFeed, setFollowingFeed] = useAtom(followingFeedAtom)
-  const router = useRouter()
-  const { data: session, status } = useSession()
-
-  useEffect(() => {
-    if (!session) return
-    const isFollowed = followingFeed.some((feed) => feed.id === item.id)
-    setIsFollowed(isFollowed)
-  }, [followingFeed])
-
-  const handleClick = async () => {
-    if (!session) {
-      router.push('/login')
-      return
-    }
-    setIsFollowed(!isFollowed)
-    if (isFollowed) {
-      const { result, error } = await deleteFollowFeed({ userId: session.user.id, feedId: item.id })
-      if (error) {
-        alert('エラーが発生しました')
-        return
-      }
-      return
-    }
-    const { result, error } = await createFollowFeed({ userId: session.user.id, feedId: item.id })
-    if (error) {
-      alert('エラーが発生しました')
-      return
-    }
-  }
   return (
     <div className="border border-[#afafaf] rounded p-3 bg-[white]">
       <div>
@@ -62,14 +26,7 @@ const FeedItem = ({
         {item._count.FollowFeed}
       </div>
       <div className="text-right">
-        <button
-          onClick={handleClick}
-          className={` px-3 py-1 rounded-xl duration-200 border border-[#afafaf] transition-all ${
-            isFollowed ? 'bg-primary text-[white]' : 'hover:bg-lightGray'
-          }`}
-        >
-          {isFollowed ? 'フォロー中' : 'フォロー'}
-        </button>
+        <FollowFeedButton feedId={item.id} />
       </div>
     </div>
   )
