+++ components/feeds/list/FeedLists.tsx
@@ -1,11 +1,7 @@
 'use client'
 
-import { useEffect } from 'react'
 import FeedItem from './FeedItem'
-import { Feed, FollowFeed, User } from '@prisma/client'
-import { useSession } from 'next-auth/react'
-import { useAtom } from 'jotai'
-import { followingFeedAtom } from '@/jotai/followingFeedAtom'
+import { Feed } from '@prisma/client'
 
 const FeedLists = ({
   feeds
@@ -16,19 +12,6 @@ const FeedLists = ({
     }
   })[]
 }) => {
-  const { data: session, status } = useSession()
-  const [followingFeed, setFollowingFeed] = useAtom(followingFeedAtom)
-  useEffect(() => {
-    if (!session) return
-    const fetchFeeds = async () => {
-      const res = await fetch(`/api/feeds/following/${session.user.id}`)
-      if (!res.ok) return
-      const data = (await res.json()) as { result: FollowFeed & { feed: Feed }[] }
-      const followingFeeds = data.result.map((item) => item.feed)
-      setFollowingFeed(followingFeeds)
-    }
-    fetchFeeds()
-  }, [session])
   return (
     <div className="flex flex-col gap-3 mt-5">
       {feeds.map((item) => {
